requirements.txt
langgraph>=0.2.52
langchain>=0.3.20
langchain-anthropic>=0.3.0
langchain-community>=0.3.8
tavily-python>=0.5.0
pydantic>=2.9.2
python-dotenv>=1.0.1